fix a bug that broke hiding and disabling system apps
The old method would always return false which would cause disabling or
hiding apps to work at first. After disabledApps or hiddenApps was
reloaded all the system apps would get flagged as installed apps in the
AppAdapter class. This normally wouldn't matter because root is required
to hide or disable apps, but whenever a check ran to compare the AppInfo
strings they wouldn't match and the UI would reflect that.